Initial commit of example twitter application using the repository pattern with use cases and model entities to create an app which is completely decoupled from any datasorce.
This is an example app I am working on while I learn how to implement these various patterns.
It will include a vagrant and docker setup once done. The view objects will be handles by
decorators (presenter) and there will be a couple examples of repositories as an example
of how to switch in a repository for whatever underlying database of your choice.

There is already quite a bit code in this initial commit as I had previously had this project
in a private bitbucket repo.